requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
autopep8==2.0.1
babel==2.17.0
backcall==0.2.0
basemap==2.0.0
basemap_data==2.0.0
beautifulsoup4==4.11.1
black==26.1.0
bleach==5.0.1
bleak==0.22.3
bleak-winrt==1.2.0
blinker==1.7.0
bs4==0.0.2
certifi==2025.8.3
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cvzone==1.6.1
cx-Oracle==8.3.0
cycler==0.11.0
dask==2025.4.1
dataclasses-json==0.6.7
DateTime==6.0
debugpy==1.6.5
decorator==5.1.1
deepmerge==2.0
defusedxml==0.7.1
distlib==0.3.8
distributed==2025.4.1
distro==1.9.0
Django==4.1.5
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.10.0
fastjsonschema==2.16.2
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.2
flask-cors==6.0.2
Flask-SQLAlchemy==0.6
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.1
# Editable Git install with no remote (Gun-Dection==0.1)-e c:\project\cnn
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.4
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_ai==2.31.4
jupyter_ai_chatgpt==0.5.0
jupyter_ai_magics==2.31.4
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.15.0
jupyter_server_terminals==0.4.4
jupyterlab==4.4.1
jupyterlab-pygments==0.2.2
jupyterlab-ui-profiler==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_rtc==0.0.5
jupyterlab_server==2.27.3
kaggle==1.8.3
kagglehub==0.4.1
kagglesdk==0.1.15
kiwisolver==1.4.4
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.37
librt==0.7.8
locket==1.0.0
MarkupSafe==2.1.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.6
mediapipe==0.10.9
memory-profiler==0.61.0
mistune==2.0.4
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy==1.19.1
mypy_extensions==1.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.6.1
nltk==3.9.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.4
openai==1.36.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.1
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.2
pathspec==1.0.3
patsy==1.0.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.1.0
pluggy==1.5.0
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.36
propcache==0.3.1
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.11
pure-eval==0.2.2
pyarrow==23.0.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.14.0
pymongo==4.8.0
PyMySQL==1.1.2
pyparsing==3.0.9
pyproj==3.7.2
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.4
python-slugify==8.0.4
pytokens==0.3.0
pytz==2022.7.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==7.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.7
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
selenium==4.35.0
Send2Trash==1.8.3
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.3.2.post1
SQLAlchemy==2.0.40
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.14.4
sympy==1.14.0
tblib==3.1.0
tenacity==9.1.2
terminado==0.17.1
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==1.0.0
torch==2.10.0
tornado==6.2
tqdm==4.66.4
traitlets==5.8.1
trio==0.30.0
trio-websocket==0.12.2
types-requests==2.32.4.20260107
types-tqdm==4.67.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2022.7
uri-template==1.2.0
urllib3==2.5.0
uv==0.4.16
virtualenv==20.25.0
wcwidth==0.2.6
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.5
wsproto==1.2.0
WTForms==3.1.2
yarl==1.20.0
zict==3.0.0
zipp==3.21.0
zope.interface==8.2
zstandard==0.23.0
